Migrate home sidebar widget
